requirements.test.txt
pytest-homeassistant-custom-component==0.12.53
pytest-asyncio
